+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/roomlist/room/Invite.kt
@@ -10,11 +10,9 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Check
 import androidx.compose.material.icons.filled.Close
-import androidx.compose.material3.Button
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
@@ -33,6 +31,8 @@ import de.connect2x.messenger.compose.view.common.MessengerModal
 import de.connect2x.messenger.compose.view.common.MessengerModalContent
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
+import de.connect2x.messenger.compose.view.theme.components
+import de.connect2x.messenger.compose.view.theme.components.ThemedButton
 import de.connect2x.trixnity.messenger.viewmodel.roomlist.RoomListElementViewModel
 
 interface InviteRoomListElement {
@@ -93,18 +93,19 @@ class InviteRoomListElementImpl : InviteRoomListElement {
                             modifier = Modifier.height(IntrinsicSize.Min)
                         ) {
                             Box(Modifier.weight(1.0f, fill = true), contentAlignment = Alignment.Center) {
-                                Button(
+                                ThemedButton(
+                                    style = MaterialTheme.components.primaryButton,
                                     onClick = { roomListElementViewModel.rejectInvitation() },
-                                    modifier = Modifier.buttonPointerModifier().fillMaxSize(),
+                                    modifier = Modifier.fillMaxSize(),
                                 ) {
                                     Text(i18n.invitationReject())
                                 }
                             }
                             Spacer(Modifier.size(20.dp))
                             Box(Modifier.weight(1.0f, fill = true), contentAlignment = Alignment.Center) {
-                                OutlinedButton(
+                                ThemedButton(
+                                    style = MaterialTheme.components.commonButton,
                                     onClick = { roomListElementViewModel.rejectInvitationAndBlockInviter() },
-                                    modifier = Modifier.buttonPointerModifier(),
                                 ) {
                                     Text(i18n.invitationBlock(), textAlign = TextAlign.Center)
                                 }