+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/verification/RedoSelfVerificationWizard.kt
@@ -1,9 +1,12 @@
 package de.connect2x.messenger.compose.view.verification
 
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Warning
-import androidx.compose.material3.Button
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
@@ -11,7 +14,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import de.connect2x.messenger.compose.view.DI
-import de.connect2x.messenger.compose.view.buttonPointerModifier
 import de.connect2x.messenger.compose.view.common.CloseModalButton
 import de.connect2x.messenger.compose.view.common.RunningText
 import de.connect2x.messenger.compose.view.common.Wizard
@@ -19,6 +21,8 @@ import de.connect2x.messenger.compose.view.common.WizardNavigationButton.Custom
 import de.connect2x.messenger.compose.view.common.WizardStep
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
+import de.connect2x.messenger.compose.view.theme.components
+import de.connect2x.messenger.compose.view.theme.components.ThemedButton
 import de.connect2x.trixnity.messenger.viewmodel.verification.RedoSelfVerificationViewModel
 
 interface RedoSelfVerificationWizardView {
@@ -69,9 +73,9 @@ class RedoSelfVerificationWizardViewImpl : RedoSelfVerificationWizardView {
             },
             nextButton = {
                 Custom {
-                    Button(
+                    ThemedButton(
+                        style = MaterialTheme.components.primaryButton,
                         onClick = redoSelfVerificationViewModel::startSelfVerification,
-                        Modifier.buttonPointerModifier()
                     ) {
                         Text(i18n.redoSelfVerificationRedo())
                     }